requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-qdrant
langchain-huggingface
qdrant-client
sentence-transformers
# switch to new Google SDK
google-genai
